pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Devin AI <158243242+devin-ai-integration[bot]@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.12"}
psycopg = {extras = ["binary"], version = "^3.2.6"}
slack-bolt = "^1.23.0"
python-dotenv = "^1.1.0"
slack-sdk = {extras = ["socket-mode"], version = "^3.35.0"}
openai = "^1.73.0"
notion-client = "^2.0.0" # Or the latest version, check PyPI

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
